requirements.txt
# Core dependencies for Gmail API Integration
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib>=1.1.0
google-auth>=2.23.4
httplib2==0.22.0
urllib3==2.1.0

# Additional recommended dependencies
python-dateutil==2.8.2  # For date handling
pytz==2023.3.post1     # For timezone support
tqdm==4.66.1           # For progress bars
